Refactor(setting_page.dart, dialogs.dart): Dialog 클래스 분리
Dialog 클래스를 dialogs.dart로 이동시킴.
